# backend/src/main/resources/docs/api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Simple Banking System API
  description: A simple API to illustrate Banking System concepts
servers:
  - url: http://localhost:8080/
paths:
...
  /customers:
    post:
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDto'
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
      security:
        - bearerAuth: [ ]
    get:
      summary: Get all customers
      operationId: getAllCustomers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerDto'
      security:
        - bearerAuth: [ ]
  /customers/{customerId}:
    get:
      summary: Get a customer by ID
      operationId: getCustomerById
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: Customer not found
      security:
        - bearerAuth: [ ]
    put:
      summary: Update a customer
      operationId: updateCustomerById
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: Customer not found
      security:
        - bearerAuth: [ ]
    delete:
      summary: Delete a customer
      operationId: deleteCustomerById
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful response
        '404':
          description: Customer not found
      security:
        - bearerAuth: [ ]
  /customers/{identityNumber}:
    get:
      summary: Get a customer by identity number
      operationId: getCustomerByIdentityNumber
      parameters:
        - in: path
          name: identityNumber
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDto'
        '404':
          description: Customer not found
      security:
        - bearerAuth: [ ]
...
  /branches:
    post:
      summary: Create a new branch
      operationId: createBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDto'
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDto'
    get:
      summary: Get all branches
      operationId: getAllBranches
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BranchDto'
  /branches/{branchId}:
    get:
      summary: Get a branch by ID
      operationId: getBranchById
      parameters:
        - in: path
          name: branchId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDto'
        '404':
          description: Branch not found
    put:
      summary: Update a branch
      operationId: updateBranchById
      parameters:
        - in: path
          name: branchId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDto'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDto'
        '404':
          description: Branch not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CustomerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        identityNumber:
          type: string
        phoneNumber:
          type: string
        alternateNumber:
          type: string
        emailAddress:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountDto'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardDto'
...
    AddressDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        street:
          type: string
        city:
          type: string
        zipCode:
          type: string
        province:
          type: string
        country:
          type: string
    AccountDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountNumber:
          type: string
        balance:
          type: number
          format: double
        accountType:
          type: string
        customer:
          $ref: '#/components/schemas/CustomerDto'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDto'
    BranchDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        code:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
    TransactionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        amount:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
        account:
          $ref: '#/components/schemas/AccountDto'
    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardNumber:
          type: string
        ccv:
          type: integer
          format: int32
        expirationDate:
          type: string
          format: date
        customer:
          $ref: '#/components/schemas/CustomerDto'
        status:
          type: string
...